[hw5] make thread run interleaved successfully

+++ lab5/include/memAlloc.h
@@ -35,7 +35,7 @@ void list_remove(list_node *entry);
 bool check_list(int index);
 int mem_init();
 void* my_alloc(int size);
-void my_free(int addr);
+void my_free(unsigned long addr);
 void* dy_alloc(int size);
 void mem_status_dump();
 void dy_mem_status_dump();
